+++ src/app/core/components/header/header.component.scss
@@ -1,4 +1,4 @@
-@import "../../shared/constants";
+@import "../../../shared/constants";
 
 :host {
   display: flex;
@@ -24,7 +24,7 @@
   display: flex;
   width: 50px;
   height: 50px;
-  background: center / contain no-repeat url(../../../assets/youtube-icon.svg);
+  background: center / contain no-repeat url(../../../../assets/youtube-icon.svg);
   margin-left: auto;
 }
