+++ src/app/core/core.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HeaderComponent } from './components/header/header.component';
+import { FormsModule } from '@angular/forms';
+import { SharedModule } from '../shared/shared.module';
+
+@NgModule({
+  declarations: [HeaderComponent],
+  imports: [
+    CommonModule,
+    FormsModule,
+    SharedModule,
+  ],
+  exports: [HeaderComponent]
+})
+export class CoreModule { }
