+++ src/app/youtube/components/filter/filter.component.scss
@@ -1,4 +1,4 @@
-@import "../../shared/constants";
+@import "../../../shared/constants";
 
 .filter {
   display: flex;
