+++ src/app/youtube/components/search-result-item/search-result-item.component.scss
@@ -1,4 +1,4 @@
-@import "../../shared/constants";
+@import "../../../shared/constants";
 
 .card {
   width: 260px;
