+++ src/app/youtube/components/search-results/search-results.component.scss
